feat (worker): worker auto-update
